Splitt komponent
